move ops classes to their own files
